+++ tests/Bundle/Controller/OidcCallbackTest.php
@@ -180,8 +180,6 @@ class OidcCallbackTest extends SymfonyTestCase
 
     public function test_gets_id_token_and_signs_up(): void
     {
-        $eventDispatcher = TestEventDispatcher::enable($this->container);
-
         [
             'privateKeyPem' => $privateKeyPem,
             'jwks' => $jwks
@@ -227,7 +225,7 @@ class OidcCallbackTest extends SymfonyTestCase
         $this->assertSame('test_client_id', $parsedBody['client_id']);
         $this->assertSame('test_client_secret', $parsedBody['client_secret']);
 
-        $event = $eventDispatcher->getFirstEvent(UserSignedUpEvent::class);
+        $event = $this->getEd()->getFirstEvent(UserSignedUpEvent::class);
         $this->assertSame('user123', $event->getUser()->oidc_sub);
     }