+++ tests/Unit/Auth/Oidc/OidcAuthTest.php
@@ -34,7 +34,7 @@ class OidcAuthTest extends SymfonyTestCase
         $oidc = $this->getOidcAuth();
         $requestNotLoggedIn = Request::create('/');
         $requestNotLoggedIn->setSession($this->createMock(SessionInterface::class));
-        $this->assertFalse($oidc->check($requestNotLoggedIn));
+        $this->assertFalse($oidc->me($requestNotLoggedIn));
     }
 
     public function test_check_has_session_but_wrong_user_id(): void
@@ -49,7 +49,7 @@ class OidcAuthTest extends SymfonyTestCase
         $requestNotLoggedIn = Request::create('/');
         $requestNotLoggedIn->setSession($session);
 
-        $this->assertFalse($oidc->check($requestNotLoggedIn));
+        $this->assertFalse($oidc->me($requestNotLoggedIn));
     }
 
     public function test_check_logged_in(): void
@@ -68,7 +68,7 @@ class OidcAuthTest extends SymfonyTestCase
         $requestNotLoggedIn = Request::create('/');
         $requestNotLoggedIn->setSession($session);
 
-        $user = $oidc->check($requestNotLoggedIn);
+        $user = $oidc->me($requestNotLoggedIn);
         $this->assertInstanceOf(AuthUser::class, $user);
 
         $this->assertSame($oidcUser->getId(), $user->id);