+++ tests/Unit/Billing/BillingFakeSymfonyTest.php
@@ -3,7 +3,6 @@
 namespace Hyvor\Internal\Tests\Unit\Billing;
 
 use Hyvor\Internal\Billing\BillingFake;
-use Hyvor\Internal\Billing\Dto\LicensesCollection;
 use Hyvor\Internal\Billing\License\BlogsLicense;
 use Hyvor\Internal\Tests\SymfonyTestCase;
 
@@ -12,7 +11,7 @@ class BillingFakeSymfonyTest extends SymfonyTestCase
 
     use BillingFakeTestTrait;
 
-    protected function enable(?BlogsLicense $license = null, LicensesCollection|callable|null $licenses = null): void
+    protected function enable(?BlogsLicense $license = null, array|callable|null $licenses = null): void
     {
         BillingFake::enableForSymfony($this->container, $license, $licenses);
     }