+++ tests/Unit/Billing/BillingFakeTestTrait.php
@@ -4,26 +4,29 @@ namespace Hyvor\Internal\Tests\Unit\Billing;
 
 use Hyvor\Internal\Billing\BillingFake;
 use Hyvor\Internal\Billing\BillingInterface;
-use Hyvor\Internal\Billing\Dto\LicenseOf;
-use Hyvor\Internal\Billing\Dto\LicensesCollection;
 use Hyvor\Internal\Billing\License\BlogsLicense;
+use Hyvor\Internal\Billing\License\Resolved\ResolvedLicense;
+use Hyvor\Internal\Billing\License\Resolved\ResolvedLicenseType;
 use Hyvor\Internal\Component\Component;
 use PHPUnit\Framework\Attributes\TestWith;
 
 trait BillingFakeTestTrait
 {
 
+    /**
+     * @param array<int, ResolvedLicense>|callable|null $licenses
+     */
     protected abstract function enable(
         ?BlogsLicense $license = null,
-        LicensesCollection|callable|null $licenses = null
+        array|callable|null $licenses = null
     ): void;
 
     public function testEnableAndLicense(): void
     {
-        $this->enable(license: new BlogsLicense);
+        $this->enable(license: BlogsLicense::trial());
         $fake = $this->getContainer()->get(BillingInterface::class);
         $this->assertInstanceOf(BillingFake::class, $fake);
-        $this->assertInstanceOf(BlogsLicense::class, $fake->license(1, 1));
+        $this->assertInstanceOf(BlogsLicense::class, $fake->license(1));
     }
 
     public function testEnableAndLicenseNull(): void
@@ -31,56 +34,43 @@ trait BillingFakeTestTrait
         $this->enable(null);
         $fake = $this->getContainer()->get(BillingInterface::class);
         $this->assertInstanceOf(BillingFake::class, $fake);
-        $this->assertNull($fake->license(1, 1));
+
+        $license = $fake->license(1);
+        // TODO:
     }
 
     #[TestWith([true])]
     #[TestWith([false])]
     public function test_enable_and_licenses(bool $callback): void
     {
-        $license1 = new BlogsLicense;
-        $license2 = new BlogsLicense(users: 5);
+        $license1 = BlogsLicense::trial();
+        $license2 = BlogsLicense::trial();
+        $license2->users = 5;
 
-        $collection = new LicensesCollection([
-            [
-                'user_id' => 1,
-                'resource_id' => 1,
-                'license' => $license1->serialize()
-            ],
-            [
-                'user_id' => 2,
-                'resource_id' => 2,
-                'license' => $license2->serialize()
-            ],
-            [
-                'user_id' => 3,
-                'resource_id' => null,
-                'license' => null
-            ]
-        ], Component::BLOGS);
+        $data = [
+            1 => new ResolvedLicense(ResolvedLicenseType::SUBSCRIPTION, $license1),
+            2 => new ResolvedLicense(ResolvedLicenseType::ENTERPRISE_CONTRACT, $license2),
+            3 => new ResolvedLicense(ResolvedLicenseType::NONE)
+        ];
 
-        $this->enable(licenses: $callback ? fn() => $collection : $collection);
+        $this->enable(licenses: $callback ? fn() => $data : $data);
         $fake = $this->getContainer()->get(BillingInterface::class);
         $this->assertInstanceOf(BillingFake::class, $fake);
 
-        $allLicenses = $fake->licenses([
-            new LicenseOf(1, 1),
-            new LicenseOf(2, 2),
-            new LicenseOf(3, null)
-        ]);
+        $allLicenses = $fake->licenses([1, 2, 3]);
 
-        $this->assertCount(3, $allLicenses->all());
+        $this->assertCount(3, $allLicenses);
 
-        $license1 = $allLicenses->of(1, 1);
+        $license1 = $allLicenses[1];
         $this->assertInstanceOf(BlogsLicense::class, $license1);
         $this->assertEquals(2, $license1->users);
 
-        $license2 = $allLicenses->of(2, 2);
+        $license2 = $allLicenses[2];
         $this->assertInstanceOf(BlogsLicense::class, $license2);
         $this->assertEquals(5, $license2->users);
 
-        $license3 = $allLicenses->of(3, null);
-        $this->assertNull($license3);
+        $license3 = $allLicenses[3];
+        // TODO:
     }
 
 }