+++ tests/Unit/Billing/BillingLaravelTest.php
@@ -3,8 +3,6 @@
 namespace Hyvor\Internal\Tests\Unit\Billing;
 
 use Hyvor\Internal\Billing\Billing;
-use Hyvor\Internal\Billing\Dto\LicenseOf;
-use Hyvor\Internal\Billing\Dto\LicensesCollection;
 use Hyvor\Internal\Billing\SubscriptionIntent;
 use Hyvor\Internal\InternalApi\InternalApi;
 use Hyvor\Internal\Tests\LaravelTestCase;
@@ -15,8 +13,6 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 #[CoversClass(Billing::class)]
 #[CoversClass(SubscriptionIntent::class)]
-#[CoversClass(LicensesCollection::class)]
-#[CoversClass(LicenseOf::class)]
 class BillingLaravelTest extends LaravelTestCase
 {